rag_project/requirements.txt
llama-index-core
llama-index-embeddings-cohere
llama-index-llms-cohere
llama-index-llms-ollama
cohere
gradio
python-dotenv
